App is now able to connect and disconnect from light
There is a small bug present where a disconnected app will only remove
functionality once the slider is used. I assume this has something to
do with the slider activating a binded state but I have toyed around
with this for a long time and have not yet found a solutiong. Other
than this, both buttons work as expected.